+++ refinedstorage-autocrafting-api/src/main/java/com/refinedmods/refinedstorage/api/autocrafting/task/ExternalPatternInputSink.java
@@ -5,13 +5,16 @@ import com.refinedmods.refinedstorage.api.core.Action;
 import com.refinedmods.refinedstorage.api.resource.ResourceAmount;
 
 import java.util.Collection;
+import javax.annotation.Nullable;
 
 import org.apiguardian.api.API;
 
 @API(status = API.Status.STABLE, since = "2.0.0-milestone.4.12")
-@FunctionalInterface
 public interface ExternalPatternInputSink {
     Result accept(Pattern pattern, Collection<ResourceAmount> resources, Action action);
+
+    @Nullable
+    ExternalPatternInputSinkKey getKey(Pattern pattern);
 
     enum Result {
         ACCEPTED,