+++ refinedstorage-network-api/src/main/java/com/refinedmods/refinedstorage/api/network/autocrafting/PatternProvider.java
@@ -1,11 +1,13 @@
 package com.refinedmods.refinedstorage.api.network.autocrafting;
 
 import com.refinedmods.refinedstorage.api.autocrafting.status.TaskStatus;
+import com.refinedmods.refinedstorage.api.autocrafting.task.ExternalPatternInputSinkKey;
 import com.refinedmods.refinedstorage.api.autocrafting.task.StepBehavior;
 import com.refinedmods.refinedstorage.api.autocrafting.task.Task;
 import com.refinedmods.refinedstorage.api.autocrafting.task.TaskId;
 
 import java.util.List;
+import javax.annotation.Nullable;
 
 import org.apiguardian.api.API;
 
@@ -24,4 +26,7 @@ public interface PatternProvider extends PatternProviderExternalPatternInputSink
     void cancelTask(TaskId taskId);
 
     List<TaskStatus> getTaskStatuses();
+
+    @Nullable
+    ExternalPatternInputSinkKey getSinkKey();
 }