+++ refinedstorage-network-api/src/main/java/com/refinedmods/refinedstorage/api/network/autocrafting/PatternProviderExternalPatternInputSink.java
@@ -9,7 +9,6 @@ import java.util.Collection;
 import org.apiguardian.api.API;
 
 @API(status = API.Status.STABLE, since = "2.0.0-milestone.4.12")
-@FunctionalInterface
 public interface PatternProviderExternalPatternInputSink {
     ExternalPatternInputSink.Result accept(Collection<ResourceAmount> resources, Action action);
 }