+++ refinedstorage-network/src/main/java/com/refinedmods/refinedstorage/api/network/impl/node/relay/RelayOutputNetworkNode.java
@@ -2,6 +2,7 @@ package com.refinedmods.refinedstorage.api.network.impl.node.relay;
 
 import com.refinedmods.refinedstorage.api.autocrafting.status.TaskStatus;
 import com.refinedmods.refinedstorage.api.autocrafting.task.ExternalPatternInputSink;
+import com.refinedmods.refinedstorage.api.autocrafting.task.ExternalPatternInputSinkKey;
 import com.refinedmods.refinedstorage.api.autocrafting.task.Task;
 import com.refinedmods.refinedstorage.api.autocrafting.task.TaskId;
 import com.refinedmods.refinedstorage.api.core.Action;
@@ -183,5 +184,11 @@ public class RelayOutputNetworkNode extends AbstractNetworkNode
     @Override
     public ExternalPatternInputSink.Result accept(final Collection<ResourceAmount> resources, final Action action) {
         return ExternalPatternInputSink.Result.SKIPPED; // TODO(feat): relay support
+    }
+
+    @Nullable
+    @Override
+    public ExternalPatternInputSinkKey getSinkKey() {
+        return null; // TODO(feat): relay support
     }
 }