+++ refinedstorage-network/src/main/java/com/refinedmods/refinedstorage/api/network/impl/node/relay/RelayOutputPatternProvider.java
@@ -3,6 +3,7 @@ package com.refinedmods.refinedstorage.api.network.impl.node.relay;
 import com.refinedmods.refinedstorage.api.autocrafting.Pattern;
 import com.refinedmods.refinedstorage.api.autocrafting.status.TaskStatus;
 import com.refinedmods.refinedstorage.api.autocrafting.task.ExternalPatternInputSink;
+import com.refinedmods.refinedstorage.api.autocrafting.task.ExternalPatternInputSinkKey;
 import com.refinedmods.refinedstorage.api.autocrafting.task.Task;
 import com.refinedmods.refinedstorage.api.autocrafting.task.TaskId;
 import com.refinedmods.refinedstorage.api.core.Action;
@@ -131,5 +132,11 @@ class RelayOutputPatternProvider implements PatternProvider, PatternListener {
     @Override
     public ExternalPatternInputSink.Result accept(final Collection<ResourceAmount> resources, final Action action) {
         return ExternalPatternInputSink.Result.SKIPPED; // TODO(feat): relay support
+    }
+
+    @Nullable
+    @Override
+    public ExternalPatternInputSinkKey getSinkKey() {
+        return null; // TODO(feat): relay support
     }
 }