+++ CHANGELOG.md
@@ -19,6 +19,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
     highest priority first. If there are missing resources, lower priority patterns are checked.
 -   The Autocrafter now faces the block you're clicking when placing it, like the other cable blocks (like the Exporter or Importer).
 -   You can no longer cancel autocrafting tasks if there is not enough space in storage to return the intermediate task storage.
+-   The Autocrafting Monitor now shows the machine in which a resource is processing.
 
 ### Fixed